ccu profiling

+++ src/framework/next/comms/ccu/ccu_representation/reps/sync/ccu_rep_loc_wait_event.cc
@@ -52,7 +52,7 @@ std::string CcuRepLocWaitEvent::Describe()
     return Hccl::StringFormat("CcuRepLocWaitEvent=id[%u], mask[%04x]", event_.Id(), event_.mask);
 }
 
-uint16_t CcuRepLocWaitEvent::GetMask()
+uint32_t CcuRepLocWaitEvent::GetMask()
 {
     return event_.mask;
 }
